broke out countdown into different component, created dashboard container for both dashboard sizes
