+++ src/styling/styles.css
@@ -62,9 +62,25 @@
   margin: 0;
 }
 
+.desktopDash {
+  display: block;
+}
+
+.mobileDash {
+  display: none;
+}
+
 @media (max-width: 970px) {
   .custom-bg {
     background-position-x: 80%;
+  }
+
+  .desktopDash {
+    display: none;
+  }
+
+  .mobileDash {
+    display: block;
   }
 }
 @media (max-width: 430px) {
